task/common: Move mulptipart and object operation to common
